# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Wegen NFT Trait Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nfts/{walletAddress}:
    get:
      operationId: getUserNfts
      tags: [nfts]
      summary: Get user's Wegen NFTs
      parameters:
        - name: walletAddress
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User's NFTs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NftListResponse"
  /nfts/{tokenId}/apply-trait:
    post:
      operationId: applyTrait
      tags: [nfts]
      summary: Apply a trait from locker to an NFT
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyTraitBody"
      responses:
        "200":
          description: Trait applied successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyTraitResponse"
        "400":
          description: Trait not in locker or invalid
  /nfts/{tokenId}/remove-trait:
    post:
      operationId: removeTrait
      tags: [nfts]
      summary: Remove a trait from an NFT (returns to locker)
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoveTraitBody"
      responses:
        "200":
          description: Trait removed and returned to locker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoveTraitResponse"
  /nfts/{tokenId}/confirm-traits:
    post:
      operationId: confirmTraits
      tags: [nfts]
      summary: Confirm current trait loadout and push metadata on-chain
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmTraitsBody"
      responses:
        "200":
          description: Metadata updated on-chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmTraitsResponse"
        "400":
          description: No equipped traits or wallet mismatch
...
components:
  schemas:
...
    PayoutSplit:
      type: object
      properties:
        walletAddress:
          type: string
          description: Ethereum wallet address receiving funds
        percentage:
          type: number
          minimum: 0
          maximum: 100
          description: Percentage of sale proceeds (0-100)
      required:
        - walletAddress
        - percentage
    Trait:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        theme:
          type: string
          nullable: true
          description: Named collection/theme this trait belongs to (e.g. "Stoner Traits", "70s Vibes")
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        mediaType:
          type: string
          nullable: true
        priceUsd:
          type: string
          description: Fixed USD price (source of truth, as string to avoid floating point issues)
        priceEth:
          type: string
          description: Live-computed ETH equivalent, snapshotted from priceUsd on last save (as string)
        priceWei:
          type: string
          description: Price in wei
        totalSupply:
          type: integer
        remainingSupply:
          type: integer
        isActive:
          type: boolean
        rarity:
          type: string
          enum: [common, uncommon, rare, legendary]
        nftCollection:
          type: string
          nullable: true
        payoutSplits:
          type: array
          items:
            $ref: "#/components/schemas/PayoutSplit"
          description: Wallet addresses and their percentage share of sale proceeds
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - priceUsd
        - priceEth
        - priceWei
        - totalSupply
        - remainingSupply
        - isActive
        - rarity
        - payoutSplits
        - createdAt
...
    LockerItem:
      type: object
      properties:
        id:
          type: integer
        traitId:
          type: integer
        walletAddress:
          type: string
        trait:
          $ref: "#/components/schemas/Trait"
        quantity:
          type: integer
        equippedToTokenId:
          type: integer
          nullable: true
        purchasedAt:
          type: string
          format: date-time
        txHash:
          type: string
          nullable: true
      required:
        - id
        - traitId
        - walletAddress
        - trait
        - quantity
        - equippedToTokenId
        - purchasedAt
...
    NftTrait:
      type: object
      properties:
        category:
          type: string
        lockerItemId:
          type: integer
        trait:
          $ref: "#/components/schemas/Trait"
      required:
        - category
        - lockerItemId
        - trait
    WegenNft:
      type: object
      properties:
        tokenId:
          type: integer
        walletAddress:
          type: string
        name:
          type: string
        imageUrl:
          type: string
        metadataTxHash:
          type: string
          nullable: true
        metadataUpdatedAt:
          type: string
          nullable: true
        variantPack:
          type: string
          nullable: true
        equippedTraits:
          type: array
          items:
            $ref: "#/components/schemas/NftTrait"
      required:
        - tokenId
        - walletAddress
        - name
        - equippedTraits
    NftListResponse:
      type: object
      properties:
        nfts:
          type: array
          items:
            $ref: "#/components/schemas/WegenNft"
        total:
          type: integer
      required:
        - nfts
        - total
    ApplyTraitBody:
      type: object
      properties:
        lockerItemId:
          type: integer
        walletAddress:
          type: string
      required:
        - lockerItemId
        - walletAddress
    ApplyTraitResponse:
      type: object
      properties:
        success:
          type: boolean
        nft:
          $ref: "#/components/schemas/WegenNft"
        lockerItem:
          $ref: "#/components/schemas/LockerItem"
      required:
        - success
        - nft
        - lockerItem
    RemoveTraitBody:
      type: object
      properties:
        category:
          type: string
        walletAddress:
          type: string
      required:
        - category
        - walletAddress
    RemoveTraitResponse:
      type: object
      properties:
        success:
          type: boolean
        nft:
          $ref: "#/components/schemas/WegenNft"
        lockerItem:
          $ref: "#/components/schemas/LockerItem"
      required:
        - success
        - nft
        - lockerItem
    ConfirmTraitsBody:
      type: object
      properties:
        walletAddress:
          type: string
        variantPack:
          type: string
          nullable: true
      required:
        - walletAddress
    ConfirmTraitsResponse:
      type: object
      properties:
        success:
          type: boolean
        txHash:
          type: string
          description: Simulated/real on-chain transaction hash
        tokenId:
          type: integer
        traitsApplied:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              name:
                type: string
            required:
              - category
              - name
      required:
        - success
        - txHash
        - tokenId
        - traitsApplied
...